Refactor parser module to extract core grammar definitions into `core` submodule.
